+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## (in progress)
 
+### Stories
+
+* [UISACQCOMP-3](https://issues.folio.org/browse/UISACQCOMP-3) Handle import of stripes-acq-components to modules and platform
+
 ## [1.0.0](https://github.com/folio-org/ui-quick-marc/tree/v1.0.0) (2020-06-12)
 
 ### Stories